Refactor: Move toward a basic module form
The reason for this is because in future I want this to work as a web service, so the new structure is:

src/{body_writer, image_fetcher, report_parser}.py
cli/graph-report.py

In future, web/{...} will be a web service working similarly. to cli (importing from src)

As of this commit the CLI is run using `python.exe  -m report_crawler.cli.graph-report .\repor
t-filtered.html --verbose` - but I'll add a simple setup.py to make it installable.